requirements.txt
influxdb==5.3.1
speedtest-cli
gpiozero==1.5.1